bringing payment processing logic to invoice, can pay for invoice, view history

+++ backend/src/main/java/com/hotel/payments/interfaces/IPayment.java
@@ -1,9 +1,7 @@
 package com.hotel.payments.interfaces;
 
-import java.util.List;
-
 public interface IPayment {
- boolean processPayment(int invoiceId, String paymentType);
+ boolean processPayment(Long invoiceId, String paymentType);
 
 //  IInvoice generateIInvoice(List<IBilling> billings);
